[FIO internal] fiohab: remove safeguard for incorrect rpmb key
The initial check was done to cover cases in which the RPMB key is
already burned (e.g. test keys), to block the close process as that
would result in OP-TEE using the HUK-derived RPMB key instead.

This causes a few problems:
- Will not work on devices without eMMC/RPMB
- Can end up burning the wrong RPMB key (bugs in OP-TEE)
- Always cause a panic during the closing process

Remove as fusing shouldn't be blocked by the RPMB state.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>